Moved AccessionModelRequest.java, AccessionRequest.java, AccessionResponse.java classess to common - used in core and gateway
