+++ build.gradle
@@ -45,7 +45,7 @@ dependencies {
         branch = 'development'
     }
 }*/
-	compile fileTree(dir: '/data/commonjar', include: '*1.2.jar')
+	compile fileTree(dir: '/data/commonjar', include: '*1.3.jar')
 	implementation('org.springframework.boot:spring-boot-starter-web-services')
 	implementation('org.springframework.boot:spring-boot-starter-data-jpa')
 	implementation('commons-beanutils:commons-beanutils:1.9.4')
