+++ src/main/java/org/recap/controller/swagger/SharedCollectionRestController.java
@@ -9,10 +9,10 @@ import org.apache.commons.lang3.StringUtils;
 import org.recap.RecapCommonConstants;
 import org.recap.RecapConstants;
 import org.recap.controller.AbstractController;
-import org.recap.model.AccessionModelRequest;
+import org.recap.model.accession.AccessionModelRequest;
 import org.recap.model.BibItemAvailabityStatusRequest;
 import org.recap.model.ItemAvailabityStatusRequest;
-import org.recap.model.acession.AccessionResponse;
+import org.recap.model.accession.AccessionResponse;
 import org.recap.model.deaccession.DeAccessionRequest;
 import org.recap.model.transfer.TransferRequest;
 import org.recap.model.transfer.TransferResponse;
