--- src/main/java/org/recap/model/AccessionRequest.java
@@ -1,14 +1,0 @@
-package org.recap.model;
-
-import lombok.Getter;
-import lombok.Setter;
-
-/**
- * Created by chenchulakshmig on 19/10/16.
- */
-@Getter
-@Setter
-public class AccessionRequest {
-    private String itemBarcode;
-    private String customerCode;
-}
