+++ src/main/resources/application.properties
@@ -35,9 +35,6 @@ scsb.swagger.api.key=recap
 management.endpoints.web.exposure.include=*
 management.endpoint.health.show-details=always
 
-#Spring Cloud URI
-spring.cloud.config.uri=http://localhost:8888
-
 #RabbitMQ Information
 spring.rabbitmq.host=localhost
 spring.rabbitmq.port=5672
