+++ src/test/java/org/recap/controller/swagger/SharedCollectionRestControllerUT.java
@@ -8,8 +8,8 @@ import org.mockito.MockitoAnnotations;
 import org.recap.RecapCommonConstants;
 import org.recap.RecapConstants;
 import org.recap.controller.BaseControllerUT;
-import org.recap.model.AccessionModelRequest;
-import org.recap.model.AccessionRequest;
+import org.recap.model.accession.AccessionModelRequest;
+import org.recap.model.accession.AccessionRequest;
 import org.recap.model.BibItemAvailabityStatusRequest;
 import org.recap.model.ItemAvailabityStatusRequest;
 import org.recap.model.deaccession.DeAccessionItem;
