+++ src/test/java/org/recap/model/ReportsResponseUT.java
@@ -2,13 +2,14 @@ package org.recap.model;
 
 import org.junit.Test;
 import org.recap.BaseTestCase;
+import org.recap.model.reports.ReportsInstitutionForm;
 import org.recap.model.search.DeaccessionItemResultsRow;
 import org.recap.model.search.IncompleteReportResultsRow;
 
-import java.util.Arrays;
+import java.util.Collections;
 import java.util.Date;
 
-import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.*;
 
 /**
  * Created by hemalathas on 23/3/17.
@@ -18,33 +19,18 @@ public class ReportsResponseUT extends BaseTestCase {
     @Test
     public void testReportResponse(){
         ReportsResponse reportsResponse = new ReportsResponse();
-        reportsResponse.setAccessionPrivatePulCount(1);
-        reportsResponse.setAccessionPrivateCulCount(1);
-        reportsResponse.setAccessionPrivateNyplCount(1);
-        reportsResponse.setAccessionPrivatePulCount(1);
-        reportsResponse.setAccessionPrivateCulCount(1);
-        reportsResponse.setAccessionPrivateNyplCount(1);
-        reportsResponse.setAccessionPrivatePulCount(1);
-        reportsResponse.setAccessionPrivateCulCount(1);
-        reportsResponse.setAccessionPrivateNyplCount(1);
-        reportsResponse.setDeaccessionPrivatePulCount(1);
-        reportsResponse.setDeaccessionPrivateCulCount(1);
-        reportsResponse.setDeaccessionPrivateNyplCount(1);
-        reportsResponse.setDeaccessionPrivatePulCount(1);
-        reportsResponse.setDeaccessionPrivateCulCount(1);
-        reportsResponse.setDeaccessionPrivateNyplCount(1);
-        reportsResponse.setDeaccessionPrivatePulCount(1);
-        reportsResponse.setDeaccessionPrivateCulCount(1);
-        reportsResponse.setDeaccessionPrivateNyplCount(1);
-        reportsResponse.setOpenPulCgdCount(1);
-        reportsResponse.setOpenCulCgdCount(1);
-        reportsResponse.setOpenNyplCgdCount(1);
-        reportsResponse.setSharedPulCgdCount(1);
-        reportsResponse.setSharedCulCgdCount(1);
-        reportsResponse.setSharedNyplCgdCount(1);
-        reportsResponse.setPrivatePulCgdCount(1);
-        reportsResponse.setPrivateCulCgdCount(1);
-        reportsResponse.setPrivateNyplCgdCount(1);
+        ReportsInstitutionForm reportsInstitutionForm = new ReportsInstitutionForm();
+        reportsInstitutionForm.setInstitution("PUL");
+        reportsInstitutionForm.setAccessionPrivateCount(1);
+        reportsInstitutionForm.setAccessionPrivateCount(1);
+        reportsInstitutionForm.setAccessionPrivateCount(1);
+        reportsInstitutionForm.setDeaccessionPrivateCount(1);
+        reportsInstitutionForm.setDeaccessionPrivateCount(1);
+        reportsInstitutionForm.setDeaccessionPrivateCount(1);
+        reportsInstitutionForm.setOpenCgdCount(1);
+        reportsInstitutionForm.setSharedCgdCount(1);
+        reportsInstitutionForm.setPrivateCgdCount(1);
+        reportsResponse.setReportsInstitutionFormList(Collections.singletonList(reportsInstitutionForm));
         reportsResponse.setTotalRecordsCount("1");
         reportsResponse.setTotalPageCount(1);
         reportsResponse.setMessage("Success");
@@ -66,36 +52,10 @@ public class ReportsResponseUT extends BaseTestCase {
         incompleteReportResultsRow.setCreatedDate(new Date().toString());
         incompleteReportResultsRow.setCustomerCode("PB");
         incompleteReportResultsRow.setOwningInstitution("PUL");
-        reportsResponse.setDeaccessionItemResultsRows(Arrays.asList(deaccessionItemResultsRow));
-        reportsResponse.setIncompleteReportResultsRows(Arrays.asList(incompleteReportResultsRow));
+        reportsResponse.setDeaccessionItemResultsRows(Collections.singletonList(deaccessionItemResultsRow));
+        reportsResponse.setIncompleteReportResultsRows(Collections.singletonList(incompleteReportResultsRow));
 
-        assertNotNull(reportsResponse.getAccessionPrivatePulCount());
-        assertNotNull(reportsResponse.getAccessionPrivateCulCount());
-        assertNotNull(reportsResponse.getAccessionPrivateNyplCount());
-        assertNotNull(reportsResponse.getAccessionSharedPulCount());
-        assertNotNull(reportsResponse.getAccessionSharedCulCount());
-        assertNotNull(reportsResponse.getAccessionSharedNyplCount());
-        assertNotNull(reportsResponse.getAccessionOpenPulCount());
-        assertNotNull(reportsResponse.getAccessionOpenCulCount());
-        assertNotNull(reportsResponse.getAccessionOpenNyplCount());
-        assertNotNull(reportsResponse.getDeaccessionPrivatePulCount());
-        assertNotNull(reportsResponse.getDeaccessionPrivateCulCount());
-        assertNotNull(reportsResponse.getDeaccessionPrivateNyplCount());
-        assertNotNull(reportsResponse.getDeaccessionSharedPulCount());
-        assertNotNull(reportsResponse.getDeaccessionSharedCulCount());
-        assertNotNull(reportsResponse.getDeaccessionSharedNyplCount());
-        assertNotNull(reportsResponse.getDeaccessionOpenPulCount());
-        assertNotNull(reportsResponse.getDeaccessionOpenCulCount());
-        assertNotNull(reportsResponse.getDeaccessionOpenNyplCount());
-        assertNotNull(reportsResponse.getOpenPulCgdCount());
-        assertNotNull(reportsResponse.getOpenCulCgdCount());
-        assertNotNull(reportsResponse.getOpenNyplCgdCount());
-        assertNotNull(reportsResponse.getSharedPulCgdCount());
-        assertNotNull(reportsResponse.getSharedCulCgdCount());
-        assertNotNull(reportsResponse.getSharedNyplCgdCount());
-        assertNotNull(reportsResponse.getPrivatePulCgdCount());
-        assertNotNull(reportsResponse.getPrivateCulCgdCount());
-        assertNotNull(reportsResponse.getPrivateNyplCgdCount());
+        assertEquals(1, reportsResponse.getReportsInstitutionFormList().size());
         assertNotNull(reportsResponse.getTotalRecordsCount());
         assertNotNull(reportsResponse.getTotalPageCount());
         assertNotNull(reportsResponse.getMessage());
